added exe

+++ RedditWallpapers.py
@@ -1,8 +1,9 @@
 import requests
 import os
 import ctypes
+import queue
+from idna import idnadata
 import configparser
-
 
 USERAGENT = "fake_user_agent v1.0"
 extension = ""
@@ -12,6 +13,7 @@ FILEPATH = os.getcwd() + '\\images\\wallpaper'
 def get_pictures_from_subreddit(data, subreddit):
     global extension
     fileNotSaved = True
+    isImage = True
     for i in range(len(data)):
         current_post = data[i]['data']
         image_url = current_post['url']
@@ -23,23 +25,24 @@ def get_pictures_from_subreddit(data, subreddit):
             image_url += '.jpeg'
             extension = '.jpeg'
         else:
-            extension = ".jpeg"
+             isImage = False
 
-        if (fileNotSaved):
+        if (fileNotSaved and isImage):
             if os.path.exists(FILEPATH + extension):
                 os.remove(FILEPATH + extension)
-            print('downloading pictures from r/' + subreddit + '.. ' + str((i*100)//len(data)) + '%')
+            print('Attempting to download picture from r/' + subreddit + '.. ')
             # redirects = False prevents thumbnails denoting removed images from getting in
             image = requests.get(image_url, allow_redirects=False)
-            if(image.status_code == 200):
+            if (image.status_code == 200):
                 try:
                     output_filehandle = open(FILEPATH + extension, mode='bx')
                     output_filehandle.write(image.content)
                 except:
-                    pass
-            if(os.path.exists(FILEPATH + extension)):
+                    print("Failed to download image...")
+            if (os.path.exists(FILEPATH + extension)):
                 print(FILEPATH + extension + " exists!")
                 fileNotSaved = False
+
 
 def getImage():
     config = configparser.ConfigParser()
@@ -50,7 +53,7 @@ def getImage():
 
     print('Connecting to r/' + subreddit)
     url = 'https://www.reddit.com/r/' + subreddit + '/top/.json?sort=top&t=' + \
-            top + '&limit=' + str(number)
+          top + '&limit=' + str(number)
     response = requests.get(url, headers={'User-agent': USERAGENT})
 
     if not response.ok:
@@ -63,9 +66,17 @@ def getImage():
     data = response.json()['data']['children']
     get_pictures_from_subreddit(data, subreddit)
 
-def setWallpaper2():
-    print(FILEPATH + extension)
-    ctypes.windll.user32.SystemParametersInfoW(20, 0, FILEPATH + extension, 3)
 
-getImage()
-setWallpaper2()
+def setWallpaper2():
+    if('.' in (extension)):
+        print(FILEPATH + extension)
+        ctypes.windll.user32.SystemParametersInfoW(20, 0, FILEPATH + extension, 3)
+    else:
+        print("No image posts were found...\n")
+
+def main():
+    getImage()
+    setWallpaper2()
+
+if __name__.endswith('__main__'):
+    main()
